@@ -1,9 +1,8 @@
-﻿namespace Dapper110.Models
-{
-    public class City
-    {
-        public int Id { get; set; }
+﻿namespace Dapper110.Models;
 
-        public string Name { get; set; }
-    }
+public class City
+{
+    public int Id { get; set; }
+
+    public string Name { get; set; }
 }
