@@ -1,7 +1,6 @@
-﻿namespace Dapper110.Models.CustomTypeHandling
+﻿namespace Dapper110.Models.CustomTypeHandling;
+
+public class Tag
 {
-    public class Tag
-    {
-        public string Value { get; set; }
-    }
+    public string Value { get; set; }
 }
