@@ -1,11 +1,10 @@
-﻿namespace Dapper110.Models.MultipleMappings
+﻿namespace Dapper110.Models.MultipleMappings;
+
+public class Address
 {
-    public class Address
-    {
-        public string Street { get; set; }
+    public string Street { get; set; }
 
-        public string ZipCode { get; set; }
+    public string ZipCode { get; set; }
 
-        public string City { get; set; }
-    }
+    public string City { get; set; }
 }
