@@ -1,11 +1,10 @@
-﻿namespace Dapper110.Models
+﻿namespace Dapper110.Models;
+
+public class User
 {
-    public class User
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public string Email { get; set; }
-    }
+    public string Email { get; set; }
 }
