@@ -1,13 +1,10 @@
-﻿using System;
+﻿namespace Dapper110.Enums;
 
-namespace Dapper110.Enums
+[Flags]
+public enum Role
 {
-    [Flags]
-    public enum Role
-    {
-        Administrator = 1,
-        User = 2,
-        Reviewer = 4,
-        Contributor = 8
-    }
+    Administrator = 1,
+    User = 2,
+    Reviewer = 4,
+    Contributor = 8
 }
