@@ -1,15 +1,12 @@
-﻿using System.Collections.Generic;
+﻿namespace Dapper110.Models.MultipleMappings;
 
-namespace Dapper110.Models.MultipleMappings
+public class Restaurant
 {
-    public class Restaurant
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public Address Address { get; set; }
+    public Address Address { get; set; }
 
-        public IList<Review> Reviews { get; set; }
-    }
+    public IList<Review> Reviews { get; set; }
 }
