@@ -1,19 +1,16 @@
-﻿using System;
+﻿namespace Dapper110.Models;
 
-namespace Dapper110.Models
+public class Review
 {
-    public class Review
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        public int UserId { get; set; }
+    public int UserId { get; set; }
 
-        public int RestaurantId { get; set; }
+    public int RestaurantId { get; set; }
 
-        public int Rating { get; set; }
+    public int Rating { get; set; }
 
-        public string Comment { get; set; }
+    public string Comment { get; set; }
 
-        public DateTime Date { get; set; }
-    }
+    public DateTime Date { get; set; }
 }
